Add language selection to start message and update dependencies
An upgrade is done to the dependencies in build.gradle with the addition of 'javax.xml.bind:jaxb-api:2.3.0'. The database URL in application-dev.yml is switched to localhost. A significant enhancement in the start message handling by adding a language selection option in MessageController.java. A new BeansConfig.java file is created for providing auditor awareness.