Add i18n support and refactoring
Various changes have been made to the code to support internationalization (i18n). For instance, 'BeansConfig' has been updated to include message sources for languages and locale resolver. The language callback was also improved and refactored in the 'MessageService'. Plus, a new 'UserService' was added to allow more efficient handling of user states. A caching mechanism was also integrated by annotating the user state methods in 'UserService'. Finally, a new emoji package was added to the project's dependencies.