Refactor transaction and payment services
The transaction success process was improved by adding an extra step of deleting the message from the telegram bot. The payment services were also updated to properly handle exceptions and to store the payment method (PayPal or Stripe) related information in the transaction log. The TransactionLogEntity was updated to store the messageId.