Update subscription message with locale support
The code has been modified to support different locales when sending a subscription message. With this change, the number of days until the subscription expires is now sent as a parameter to a localised string. This string is then used to create the message that is sent to the user, in the language set in their profile.